requirements.txt
streamlit
PyPDF2
python-docx
nltk
sentence-transformers
faiss-cpu
numpy
google-generativeai
python-dotenv
